test(Coverage): With istanbu
I have added testing coverage and coverage checking to this project. The check coverage script has
been added to travis and to the pre-commit ghooks. I added in the webpack config fiel so that I can
build the es5 version of the project. I changed the main file in package.json to reflect this.